chore: migrate to TS useRemoveAsset
